modify and integrate Layout.jsx
