+++ public/manifest.json
@@ -1,7 +1,7 @@
 {
   "name":"Kawano Yudai's site",
   "short_name":"Kawano Yudai",
-  "vercel": "https://next-portfolio-blue.now.sh",
+  "vercel": "https://oriverk.dev",
   "start_url": "/",
   "display":"standalone",
   "background_color":"#212121",
