+++ src/components/HomeContents.jsx
@@ -12,6 +12,35 @@ export function Top() {
           <div className='topSubtitle'>B.Agr. /<wbr />JobSeeker</div>
         </div>
       </section>
+      <style jsx>{`
+        .topContainer {
+          height: 100vh;
+          position: relative;
+          background-color: #212121;
+          background-size: cover;
+          background-position: center;
+          background-image: url('/assets/adelaide2.webp');
+        }
+        .topTitleContainer {
+          color: white;
+          padding: 5%;
+          text-align: center;
+          position: absolute;
+          top: 50%;
+          left: 50%;
+          transform: translate(-50%, -50%);
+          background-color: rgba(0,0,0,0.3);
+          border-bottom: 2px double rgb(255,255,255);
+          border-radius: 2px;
+        }
+        .topTitle {
+          font-weight: 500;
+          font-size: 4rem;
+        }
+        .topSubtitle {
+          font-size: 2rem
+        }
+      `}</style>
     </>
   )
 }
@@ -39,6 +68,11 @@ export function About() {
           </Grid>
         </Grid>
       </section>
+      <style jsx>{`
+        #about p{
+          margin: 0 auto .5rem;
+        }
+      `}</style>
     </>    
   );
 }
@@ -136,6 +170,11 @@ export function Works() {
           </Grid>
         </article>
       </section>
+      <style jsx>{`
+        #works{
+          margin-bottom: 3.5rem;
+        }
+      `}</style>
     </>  
   );
 }
