+++ src/utils/noImage.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 export default function NoImage (){
   return (
-    <React.Fragment>
+    <>
       <div className='parent'>
         <div className='child'>no image</div>
       </div>
@@ -24,6 +24,6 @@ export default function NoImage (){
           font-weight:bold;
         }
       `}</style>
-    </React.Fragment>
+    </>
   )
 }
